Use Tomcat embedded server 8.5.60
Update servlet API to 3.1 which is the version supported by Tomcat 8.5.60.

Fix usage of HttpServletRequest.getParameterMap() in RequestUtil since
we can now assume that the value is a String[].

Don't assume that Servlet, JSP, and other Web APIs are provided by the
runtime since those assumptions cannot hold for an embedded server.

Remove Jetty vestiges.