Load CLI tools as beans
Allow usage of @Autowired in the CLI definitions since they are loaded
as beans.

Deprecate most of the AbstractSpringAwareCLI logic since it can now be
done via annotations.

Use ApplicationContext in GemmaCLI and SpringContextUtil so that we can
reuse the component scan logic via getBeansOfType.

Merge workerContext with cliContext as we don't need to make
distinctions between the two.